feat: 킵고잉 이용내역 controller 합치기
