+++ src/main/java/com/shinhan/memento/api/MyPageApiController.java
@@ -1,9 +1,12 @@
 package com.shinhan.memento.api;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import javax.servlet.http.HttpSession;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -12,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.shinhan.memento.common.exception.MemberException;
 import com.shinhan.memento.common.exception.MypageException;
 import com.shinhan.memento.common.response.BaseResponse;
 import com.shinhan.memento.common.response.status.BaseExceptionResponseStatus;
@@ -19,20 +23,28 @@ import com.shinhan.memento.dto.ConfirmCashRequestDTO;
 import com.shinhan.memento.dto.ConfirmCashResponseDTO;
 import com.shinhan.memento.dto.MyMatchupListResponseDTO;
 import com.shinhan.memento.dto.MyMentosListResponseDTO;
+import com.shinhan.memento.dto.MypageKeepgoingHistoryDTO;
 import com.shinhan.memento.dto.ValidateCashRequestDTO;
 import com.shinhan.memento.dto.ValidateCashResponseDTO;
+import com.shinhan.memento.model.BaseStatus;
 import com.shinhan.memento.model.Member;
+import com.shinhan.memento.model.UserType;
+import com.shinhan.memento.service.MemberKeepgoingService;
+import com.shinhan.memento.service.MemberService;
 import com.shinhan.memento.service.MyPageService;
-
+import lombok.extern.slf4j.Slf4j;
 import lombok.AllArgsConstructor;
 
+@Slf4j
 @RestController
 @AllArgsConstructor
 @RequestMapping("/api/mypage")
 public class MyPageApiController {
 	
 	private final MyPageService myPageService;
-
+	private final MemberKeepgoingService memberKeepgoingService;
+	private final MemberService memberService;
+	
 	@PostMapping("/validate-cash")
     public BaseResponse<ValidateCashResponseDTO> validateCash(@RequestBody ValidateCashRequestDTO reqDTO,HttpSession session) {
 		Member loginUser = (Member) session.getAttribute("loginUser");
@@ -74,6 +86,24 @@ public class MyPageApiController {
 		}
 		
 		return new BaseResponse<>(memberMatchUpList);
+	}
+	
+	/**
+	 * 마이페이지(킵고잉 이용내역 조회)
+	 */
+	@GetMapping("/history/keepgoing")
+	public BaseResponse<List<MypageKeepgoingHistoryDTO>> showKeepgoingHistoryByMemberId(@RequestParam int memberId) {
+		log.info("[MypageController.showKeepgoingHistory]");
+		// mentoId 로 들어온 식별자값이 db에서 유효한 사용자인지 검증
+		Map<String, Object> memberCheckParams = new HashMap<>();
+		memberCheckParams.put("memberId", memberId);
+		memberCheckParams.put("userType", String.valueOf(UserType.MENTO));
+		memberCheckParams.put("status", String.valueOf(BaseStatus.ACTIVE));
+		Member member = memberService.findMemberById(memberCheckParams);
+		if (member == null) {
+			throw new MemberException(BaseExceptionResponseStatus.CANNOT_FOUND_MEMBER);
+		}
+		return new BaseResponse<>(memberKeepgoingService.showKeepgoingHistoryByMemberId(memberId));
 	}
